refactor: apply linting suggestions
