+++ README.md
@@ -12,6 +12,7 @@
   - Lotto는 불변 객체이기 때문에 그냥 Controller에게 전달해서 출력하도록 함
 
 [//]: # (TODO LottoService의 정체성에 대해 고민해보기)
+
 1. 로또 구매 후 보관
 2. 당첨 로또 발매 후 보관
 3. 당첨 통계 발급
