+++ docs/feedback.md
@@ -54,10 +54,10 @@
 - 작은 기능부터 차근차근 테스트를 작성해 나가다 보면 빠르게 익숙해질 거예요.
 - 실패하거나 돌아가는 길도 괜찮습니다. 그 모든 흔적이 여러분의 몰입과 성장을 보여주는 소중한 발자국이 됩니다.
 - 여러분의 도전 과정을 담아주세요.
-    - 이번 미션에서 가장 오래 고민했던 부분과 시도들
-    - 실패했지만 의미 있었던 도전
-    - 다음에는 다르게 접근해보고 싶은 부분
-    - 여러분만의 고민과 성장 과정
+  - 이번 미션에서 가장 오래 고민했던 부분과 시도들
+  - 실패했지만 의미 있었던 도전
+  - 다음에는 다르게 접근해보고 싶은 부분
+  - 여러분만의 고민과 성장 과정
 - 관련 함수를 묶어 클래스를 만들고, 객체들이 협력하여 하나의 큰 기능을 수행하도록 한다.
 - 클래스와 함수에 대한 단위 테스트를 통해 의도한 대로 정확하게 작동하는 영역을 확보한다.
